fix: preserve snapshot task calendar boundary
