+++ backend/tests/unit/test_feature_store_tasks.py
@@ -23,6 +23,7 @@ from app.interfaces.tasks.feature_store_tasks import (
     _create_auto_scan_for_published_run,
     _enrich_feature_run_with_ibd_metadata,
     _fail_stale_feature_runs,
+    _is_market_trading_day,
     _repair_current_us_group_metadata,
     _upsert_feature_run_pointer,
     build_daily_snapshot,
@@ -80,13 +81,30 @@ class _NonSkippingUoW:
         return False
 
 
+def test_task_trading_day_guard_uses_market_calendar(monkeypatch):
+    calls = []
+
+    def is_trading_day(_self, market, day):
+        calls.append((market, day))
+        return True
+
+    monkeypatch.setattr(
+        "app.services.market_calendar_service.MarketCalendarService.is_trading_day",
+        is_trading_day,
+    )
+
+    target = date(2026, 2, 17)
+    assert _is_market_trading_day(target, market="US") is True
+    assert calls == [("US", target)]
+
+
 def test_build_daily_snapshot_normalizes_default_active_universe():
     fake_use_case = _FakeUseCase()
 
     with patch(
         "app.interfaces.tasks.feature_store_tasks.date"
     ) as mock_date, patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -144,7 +162,7 @@ def test_build_daily_snapshot_returns_failure_diagnostics(monkeypatch):
         lambda: _FakeUseCase(),
     )
     monkeypatch.setattr(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         lambda *_args, **_kwargs: True,
     )
 
@@ -185,7 +203,7 @@ def test_build_daily_snapshot_never_passes_legacy_dict_shape():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -215,7 +233,7 @@ def test_build_daily_snapshot_uses_default_scan_profile_when_not_provided():
     defaults = get_default_scan_profile("US")
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -247,7 +265,7 @@ def test_build_daily_snapshot_bootstrap_uses_lightweight_scan_profile():
     bootstrap_defaults = get_bootstrap_scan_profile("US")
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -286,7 +304,7 @@ def test_build_daily_snapshot_bootstrap_lifecycle_without_cache_gate_uses_full_p
     defaults = get_default_scan_profile("US")
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -321,7 +339,7 @@ def test_build_daily_snapshot_bootstrap_preserves_explicit_screener_names():
     explicit_screeners = ["setup_engine"]
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -421,7 +439,7 @@ def test_build_daily_snapshot_skip_if_published_requires_exact_signature_match()
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.interfaces.tasks.feature_store_tasks.hash_scan_signature",
@@ -496,7 +514,7 @@ def test_build_daily_snapshot_bootstrap_skip_if_published_uses_cache_only_suppor
         return ",".join(symbols)
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.interfaces.tasks.feature_store_tasks.hash_scan_signature",
@@ -563,7 +581,7 @@ def test_build_daily_snapshot_skip_if_published_repairs_requested_pointer():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -660,7 +678,7 @@ def test_build_daily_snapshot_reraises_soft_time_limit():
             raise SoftTimeLimitExceeded()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -692,7 +710,7 @@ def test_build_daily_snapshot_reraises_soft_time_limit_when_failure_activity_pub
             raise SoftTimeLimitExceeded()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -723,7 +741,7 @@ def test_build_daily_snapshot_creates_auto_scan_after_publish():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -761,7 +779,7 @@ def test_build_daily_snapshot_static_daily_mode_requires_bulk_prefetch():
         "app.utils.parallelism.os.cpu_count",
         return_value=4,
     ), patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -809,7 +827,7 @@ def test_build_daily_snapshot_static_daily_mode_keeps_setting_as_upper_bound_on_
         "app.utils.parallelism.os.cpu_count",
         return_value=16,
     ), patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -845,7 +863,7 @@ def test_build_daily_snapshot_static_daily_mode_uses_null_progress_sink():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -886,7 +904,7 @@ def test_build_daily_snapshot_bootstrap_gate_pass_wires_cache_only_without_null_
         "app.utils.parallelism.os.cpu_count",
         return_value=4,
     ), patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -964,7 +982,7 @@ def test_build_daily_snapshot_bootstrap_gate_fail_retries_without_live_fallback(
             raise Retry(message=str(exc))
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1017,7 +1035,7 @@ def test_build_daily_snapshot_bootstrap_gate_exhaustion_fails_with_coverage_repo
             raise AssertionError("retry must not be called after retry budget is exhausted")
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1576,7 +1594,7 @@ def test_build_daily_snapshot_includes_cleaned_stale_runs_metadata():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1608,7 +1626,7 @@ def test_build_daily_snapshot_enriches_published_run_metadata_after_publish():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1737,7 +1755,7 @@ def test_build_daily_snapshot_publishes_market_activity():
     completed = []
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1780,7 +1798,7 @@ def test_build_daily_snapshot_skips_disabled_market_by_default():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1817,7 +1835,7 @@ def test_build_daily_snapshot_allows_disabled_market_when_runtime_gate_is_ignore
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
@@ -1857,7 +1875,7 @@ def test_build_daily_snapshot_does_not_ignore_market_gate_outside_static_mode():
     fake_use_case = _FakeUseCase()
 
     with patch(
-        "app.use_cases.feature_store.build_daily_snapshot._is_us_trading_day",
+        "app.interfaces.tasks.feature_store_tasks._is_market_trading_day",
         return_value=True,
     ), patch(
         "app.wiring.bootstrap.get_build_daily_snapshot_use_case",
